requirements.txt
aiofiles==23.2.1
aiohappyeyeballs==2.4.4
aiohttp==3.11.12
aiosignal==1.3.2
aiosqlite==0.21.0
altgraph==0.17.4
annotated-types==0.7.0
anyio==4.8.0
asyncer==0.0.7
asyncpg==0.30.0
attrs==25.1.0
bidict==0.23.1
bottle==0.13.2
certifi==2025.1.31
cffi==1.17.1
chainlit==2.1.2
charset-normalizer==3.4.1
chevron==0.14.0
click==8.1.8
clr_loader==0.2.7.post0
colorama==0.4.6
cryptography==44.0.0
dataclasses-json==0.6.7
Deprecated==1.2.18
distro==1.9.0
fastapi==0.115.8
filetype==1.2.0
frozenlist==1.5.0
googleapis-common-protos==1.66.0
greenlet==3.1.1
grpcio==1.70.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
httpx-sse==0.4.0
idna==3.10
importlib_metadata==8.5.0
jiter==0.8.2
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.17
langchain-community==0.3.16
langchain-core==0.3.34
langchain-text-splitters==0.3.6
langsmith==0.3.6
Lazify==0.4.0
literalai==0.1.103
marshmallow==3.26.1
multidict==6.1.0
mypy-extensions==1.0.0
nest-asyncio==1.6.0
numpy==2.2.3
openai==1.61.0
opentelemetry-api==1.29.0
opentelemetry-exporter-otlp==1.29.0
opentelemetry-exporter-otlp-proto-common==1.29.0
opentelemetry-exporter-otlp-proto-grpc==1.29.0
opentelemetry-exporter-otlp-proto-http==1.29.0
opentelemetry-instrumentation==0.50b0
opentelemetry-proto==1.29.0
opentelemetry-sdk==1.29.0
opentelemetry-semantic-conventions==0.50b0
orjson==3.10.15
packaging==24.2
pefile==2023.2.7
propcache==0.2.1
protobuf==5.29.3
proxy_tools==0.1.0
pycparser==2.22
pydantic==2.10.6
pydantic-settings==2.7.1
pydantic_core==2.27.2
pyinstaller==6.12.0
pyinstaller-hooks-contrib==2025.1
PyJWT==2.10.1
python-dotenv==1.0.1
python-engineio==4.11.2
python-multipart==0.0.18
python-socketio==5.12.1
pythonnet==3.0.5
pywebview==5.4
pywin32-ctypes==0.2.3
PyYAML==6.0.2
requests==2.32.3
requests-toolbelt==1.0.0
setuptools==75.8.2
simple-websocket==1.1.0
sniffio==1.3.1
SQLAlchemy==2.0.38
starlette==0.41.3
syncer==2.0.3
tenacity==9.0.0
tomli==2.2.1
tqdm==4.67.1
typing-inspect==0.9.0
typing_extensions==4.12.2
uptrace==1.29.0
urllib3==2.3.0
uvicorn==0.34.0
watchfiles==0.20.0
wrapt==1.17.2
wsproto==1.2.0
yarl==1.18.3
zipp==3.21.0
zstandard==0.23.0
